refactor(package): move camera into utils
