+++ app/src/main/java/com/example/ezvault/utils/camera/CameraAction.java
@@ -1,4 +1,4 @@
-package com.example.ezvault.camera;
+package com.example.ezvault.utils.camera;
 
 import androidx.activity.ComponentActivity;
 import androidx.activity.result.ActivityResultRegistry;
