+++ app/src/main/java/com/example/ezvault/view/EditItemDetails.java
@@ -41,7 +41,7 @@ import android.widget.Toast;
 import com.example.ezvault.viewmodel.SelectedItemViewModel;
 import com.example.ezvault.view.adapter.PhotoAdapter;
 import com.example.ezvault.R;
-import com.example.ezvault.camera.GalleryAction;
+import com.example.ezvault.utils.camera.GalleryAction;
 import com.example.ezvault.data.database.FirebaseBundle;
 import com.example.ezvault.data.database.ImageDAO;
 import com.example.ezvault.data.database.ItemDAO;
